% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_lcmv}
\alias{demo_lcmv}
\title{Demo data set}
\format{
161287 observations (cells) and 30 variables (markers, scaled to 0-1023).
\describe{
  \item{FSC.A}{Forward scatter area}
  \item{FSC.H}{Forward scatter height}
  \item{SSC.A}{Side scatter area}
  \item{SSC.H}{Side scatter height}
  \item{fluorescence markers}{TCRb, CD3, NK11, MHCII, CD24, CD11c, ICOS, CD11b, CD451, CD19, CD117, Ly6G, CD138, Ly6C, CD901, CD25, CD452, CD68, F480, CD357, CD172, CD317, TCRgd, CD8, CD4, IgD}
}
}
\usage{
data(demo_lcmv)
}
\description{
Demo data set for the PICtR workflow containing 10 percent randomly sampled cells from the spleen of one LCMV infected mouse (day 7, Figure 5 and 6).
Data was analysed using high parametric spectral flow cytometry (n=26 markers).
}
\keyword{datasets}
