% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_obj}
\alias{demo_obj}
\title{Demo Seurat object}
\format{
A Seurat object with 161287 samples (cells) and 31 Features. n = 5000 cells were sketched.
\describe{
  \item{meta.data}{Contains all scatter and fluorescence markers,
  the calculated FSC.A/FSC.H ratio,
  the classification into ratio_high/ratio_low cells (ratio_anno),
  the clustering solutions for sketched cells with resolutions of 0.8 and 1 (sketch_snn_res),
  and predicted cluster labels for all cells (clusters_predicted_obj).}
  \item{assay "FACS"}{The Seurat assay for n = 161287 cells.}
  \item{assay "sketch"}{The Seurat assay for n = 5000 sketched cells.}
  \item{reductions}{Calculated dimensional reductions (PCA, UMAP).}
  \item{misc}{List containing the selected clusters representing physically interacting cells.}
}
}
\usage{
data(demo_obj)
}
\description{
Demo Seurat object based on the demo data set for the spleen of one LCMV infected mouse at day 7.
}
\keyword{datasets}
