% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cytometry_data.R
\name{import_cytometry_data}
\alias{import_cytometry_data}
\title{Data Import Helper}
\usage{
import_cytometry_data(
  path,
  filetype = "csv",
  sep = ",",
  verbose = T,
  downsample = NULL
)
}
\arguments{
\item{path}{The path to the directory containing the transformed and gated cytometry data files (string).}

\item{filetype}{The file type of the cytometry data files (string, default: "csv").}

\item{sep}{The field separator character used in the cytometry data files. See \code{\link[data.table]{fread}} (string, default: `,`).}

\item{verbose}{Verbosity (Boolean, default: `TRUE`).}

\item{downsample}{Optional random down-sampling of each individual file before they are combined into one data frame. If provided, the proportion of events that should be selected (numerical, default: `NULL`).}
}
\value{
A `data.table`.
}
\description{
Data Import Helper
}
