% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{MEM_heatmap}
\alias{MEM_heatmap}
\title{Marker Enrichment Modeling (MEM) Heat map}
\usage{
MEM_heatmap(
  obj,
  markers = c(),
  cluster_col = "seurat_clusters",
  cols = pals::coolwarm(100),
  heatmap_name = "MEM enrichment score",
  heatmap_column_title = "marker",
  heatmap_row_title = "cluster",
  scale_width = 2.2,
  scale_height = 5
)
}
\arguments{
\item{obj}{The Seurat object.}

\item{markers}{meta.data columns with features that should be plotted in the heat map and the clustering solution.}

\item{cluster_col}{Character string specifying the column that contains the clustering solution.}

\item{cols}{Color palette.}

\item{heatmap_name}{Title of the heat map.}

\item{heatmap_column_title}{Title for the columns of the heat map.}

\item{heatmap_row_title}{Title for the rows of the heat map.}

\item{scale_width}{Scaling factor for the width of the heat map in relation to the number of columns.}

\item{scale_height}{Scaling factor for the height of the heat map in relation to the number of rows.}
}
\value{
MEM heat map
}
\description{
Marker Enrichment Modeling (MEM) Heat map
}
