% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_data.R
\name{predict_data}
\alias{predict_data}
\title{Cluster label prediction}
\usage{
predict_data(
  obj = obj,
  data_query = query,
  ref_clusters = NULL,
  FSC.A = "FSC.A",
  FSC.H = "FSC.H",
  pred_name = "clusters_predicted",
  assay_ref = NULL,
  assay_query = NULL,
  chunk_size = 1e+06,
  return_obj = TRUE
)
}
\arguments{
\item{obj}{The Seurat object.}

\item{data_query}{A data frame or a Seurat object with the cells whose labels should be predicted.}

\item{ref_clusters}{The meta.data column with the reference cluster labels.}

\item{FSC.A}{The name of the column containing the FSC.A scatter parameter.}

\item{FSC.H}{The name of the column containing the FSC.H scatter parameter.}

\item{pred_name}{The name of the meta.data column for predicted cluster labels (character vector).}

\item{assay_ref}{The name of the Seurat assay which was used to calculate the reference cluster labels.}

\item{assay_query}{The name of the Seurat assay containing cells whose labels should be predicted. Only if the query is provided as a Seurat object.}

\item{chunk_size}{Chunk size for the prediction progress for verbose output to standard out.}

\item{return_obj}{Boolean. Add the predicted cluster labels to the Seurat object? Only if the query is provided as a data frame.}
}
\value{
Seurat object or data frame containing the predicted cluster labels.
}
\description{
Predicts the cluster labels for a reference data set to a query data set using \code{\link[MASS]{lda}}.
}
\references{
Venables, W. N. & Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth Edition. Springer, New York. ISBN 0-387-95457-0

Ripley, B. D. (1996) Pattern Recognition and Neural Networks. Cambridge University Press.
}
