% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{save.obj}
\alias{save.obj}
\title{Save Seurat object}
\usage{
save.obj(object, path = output_obj_path, file)
}
\arguments{
\item{object}{The Seurat Object.}

\item{path}{Directory path.}

\item{file}{File name.}
}
\value{
None
}
\description{
Wrapper function for saving Seurat objects
}
