% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{save.plot}
\alias{save.plot}
\title{Save plot}
\usage{
save.plot(plot, path = output_plot_path, file, dim_h = 7, dim_w = 5)
}
\arguments{
\item{plot}{The plot.}

\item{path}{Directory path.}

\item{file}{The file name.}

\item{dim_h}{Height parameter for \code{\link[grDevices]{pdf}}}

\item{dim_w}{Width parameter for \code{\link[grDevices]{pdf}}}
}
\value{
None
}
\description{
Wrapper function for saving plots as a pdf file.
}
