% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_dbt.R
\name{select_dbt}
\alias{select_dbt}
\title{Selection of clusters containing physically interacting cells.}
\usage{
select_dbt(
  obj,
  clusters = "seurat_clusters",
  ratio = "ratio_anno",
  ratio_high = "Ratio_high",
  assay = "FACS",
  quantile = 0.8,
  selected_clusters = "doublet_clusters"
)
}
\arguments{
\item{obj}{The Seurat object.}

\item{clusters}{The meta.data column containing the cluster labels.}

\item{ratio}{The meta.data column with the classification of cells into ratio_high and ratio_low using the FSC ratio (FSC.A/FSC.H) and Otsu's thresholding method.}

\item{ratio_high}{The character string that indicates cells with a FSC ratio (FSC.A/FSC.H) above the threshold determined with calculateThreshold() within the meta.data column.}

\item{assay}{The Seurat assay containing FACS data.}

\item{quantile}{The desired percentile cutoff above which clusters are classified as physically interacting cell clusters (Default 0.8, meaning top 20 percent of clusters are chosen).}

\item{selected_clusters}{Character vector for the misc slot in the Seurat object that will contain the cluster numbers of the selected physically interacting cell clusters.}
}
\value{
Seurat object
}
\description{
Selects the clusters that belong to a given percentile of the clusters with the largest physically interacting cells proportion based on FSC ratio thresholding.
}
