% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_function.R
\name{sketch_wrapper}
\alias{sketch_wrapper}
\title{PICtR workflow wrapper}
\usage{
sketch_wrapper(
  channel = channel,
  meta_data = NULL,
  assay = "FACS",
  FSC.A = "FSC.A",
  FSC.H = "FSC.H",
  n_sketch_cells = 50000,
  resolution = c(0.5, 1, 2, 3, 4),
  obj_name = "obj_sketched_non_projected",
  group_by = NULL,
  verbose = TRUE,
  BPcell_dir = NULL,
  ratio = TRUE,
  working_dir = getwd()
)
}
\arguments{
\item{channel}{A data frame with the dimensionality cells x flow cytometry parameters.}

\item{meta_data}{A data frame with meta_data for every cell.}

\item{assay}{A character string with the name of the assay.}

\item{FSC.A}{The name of the column containing the FSC.A scatter parameter.}

\item{FSC.H}{The name of the column containing the FSC.H scatter parameter.}

\item{n_sketch_cells}{The number of cells to be subsampled by \code{\link[Seurat]{SketchData}}.}

\item{resolution}{A character vector with the desired resolutions for clustering.}

\item{obj_name}{The name of the Seurat object.}

\item{group_by}{Optional grouping parameter to calculate the FSC.A/FSC.H Otsu threshold for.}

\item{verbose}{Verbosity.}

\item{BPcell_dir}{Optional directory with the counts matrix for \code{\link[BPCells]{open_matrix_dir}}.}

\item{ratio}{Should the ratio be calculated?}

\item{working_dir}{Working directory as a character string.}
}
\value{
Seurat object.
}
\description{
Characterizes cells according to the Otsu threshold, samples a representative subset of cells using \code{\link[Seurat]{SketchData}} and runs the standard Seurat analysis workflow
}
