% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{umap_rasterized}
\alias{umap_rasterized}
\title{Rasterized UMAP}
\usage{
umap_rasterized(
  data = obj,
  group.by = "seurat_clusters",
  raster.dpi = 500,
  label = TRUE,
  cols = pals::tol.rainbow(70),
  umap1 = "umap_1",
  umap2 = "umap_2",
  reduction = "umap",
  raster = TRUE
)
}
\arguments{
\item{data}{Seurat object.}

\item{group.by}{meta.data column to group the dimensional reduction plot by, for exmaple a clustering solution.}

\item{raster.dpi}{Pixel resolution (numeric; default 500)}

\item{label}{Plot labels?}

\item{cols}{Color palette.}

\item{umap1}{First UMAP dimension as found in the meta.data.}

\item{umap2}{Second UMAP dimension as found in the meta.data.}

\item{reduction}{Whcih reduction to use.}

\item{raster}{Rasterize the plot?}
}
\value{
Plot
}
\description{
Wrapper function to plot a dimensional reduction plot colored by clusters.
}
