% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{wrapper_for_plots}
\alias{wrapper_for_plots}
\title{Plot wrapper function.}
\usage{
wrapper_for_plots(
  obj = obj,
  feature_plot = TRUE,
  cluster_plot = TRUE,
  meta_list = list("ratio_anno"),
  cluster_handle = "sketch_snn_res",
  feature_plot_colors = pals::parula(1000),
  ratio_plot_color = c(Ratio_low = "dodgerblue2", Ratio_high = "gold2"),
  reduction = "umap",
  alpha = 1,
  raster = F,
  label_size = 3,
  label_box = FALSE,
  assay = "sketch"
)
}
\arguments{
\item{obj}{The Seurat object.}

\item{feature_plot}{Boolean. TRUE indicates that UMAP is colored for all features (see \code{\link[Seurat]{FeaturePlot}})}

\item{cluster_plot}{Boolean. TRUE indicates that UMAP is colored by the different cluster resolutions stored with the cluster_handle.}

\item{meta_list}{List of meta.data columns to color \code{\link[Seurat]{DimPlot}} by.
Can be both numeric to generate \code{\link[Seurat]{FeaturePlot}} or class character or factor for \code{\link[Seurat]{DimPlot}}.}

\item{cluster_handle}{Prefix for the clustering solutions in the meta.data slot.}

\item{feature_plot_colors}{Color palette for \code{\link[Seurat]{FeaturePlot}}.}

\item{ratio_plot_color}{Color palette for \code{\link{ratio_cluster_plot}}.}

\item{reduction}{Reduction to use for plotting, for example UMAP.}

\item{alpha}{Alpha value for plotting.}

\item{raster}{Convert points to raster format. If TRUE plot is rasterized to raster.dpi=c(512, 512). Requires ggrastr.}

\item{label_size}{Size of the labels plotted within the embedding.}

\item{label_box}{Plot boxes around labels in the color of the cluster.}

\item{assay}{The Seurat assay (default FACS).}
}
\value{
A list with the requested plots.
}
\description{
A wrapper function for plots that are often used during exploratory analysis within the Seurat framework (\url{https://satijalab.org/seurat/}).
Includes \code{\link[Seurat]{FeaturePlot}}, \code{\link[Seurat]{DimPlot}} for different parameters, and \code{\link{ratio_cluster_plot}}.
}
