% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateThreshold.R
\name{calculateThreshold}
\alias{calculateThreshold}
\title{Otsu Threshold}
\usage{
calculateThreshold(hist)
}
\arguments{
\item{hist}{A histogram to calculate the Otsu threshold for, for example created by \code{\link[graphics]{hist}}}
}
\value{
Numerical threshold value
}
\description{
Calculates the Otsu threshold for a given histogram.
}
