% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateThreshold.R
\name{calculateThreshold}
\alias{calculateThreshold}
\title{Threshold calculation}
\usage{
calculateThreshold(data, method = "otsu", breaks = 2000, seed = 42)
}
\arguments{
\item{data}{The data to calculate a threshold for. Must be non-negative integers for autothresholdr methods, see below.}

\item{method}{The method used for thresholding. One of: "otsu", "triangle", "kmeans" or methods from the autothresholdr package, see \code{\link[autothresholdr]{auto_thresh}} ("IJDefault", "Huang", "Huang2", "Intermodes", "IsoData", "Li", "MaxEntropy", "Mean", "MinErrorI", "Minimum", "Moments", "Otsu", "Percentile", "RenyiEntropy", "Shanbhag", "Triangle" and "Yen"). For autothresholdr methods, the data will be forced to integers (Default: "otsu").}

\item{breaks}{Number of histogram breaks for methods that rely on histograms. Should be >100 (Default: 2000).}

\item{seed}{Seed for reproducibility (Default: 42).}
}
\value{
Numerical threshold value
}
\description{
Calculates a threshold based on different methods.
}
\examples{
data("demo_lcmv")
demo_lcmv$ratio <- demo_lcmv$FSC.A/demo_lcmv$FSC.H
threshold <- calculateThreshold(demo_lcmv$ratio, method = "otsu", breaks = 2000)

}
\references{
N. Otsu, "A Threshold Selection Method from Gray-Level Histograms," in IEEE Transactions on Systems, Man, and Cybernetics, vol. 9, no. 1, pp. 62-66, Jan. 1979, doi: \url{10.1109/TSMC.1979.4310076}.

Zack GW, Rogers WE, Latt SA. Automatic measurement of sister chromatid exchange frequency. Journal of Histochemistry & Cytochemistry. 1977;25(7):741-753. doi: \url{10.1177/25.7.70454}.

Hartigan, J. A., & Wong, M. A. (1979). Algorithm AS 136: A K-Means Clustering Algorithm. Journal of the Royal Statistical Society. Series C (Applied Statistics), 28(1), 100–108. \url{https://doi.org/10.2307/2346830}.

G. Landini, D.A. Randell, S. Fouad, and A. Galton. Automatic thresholding from the gradients of region boundaries. Journal of Microscopy, 265(2), 185-195.
}
