% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PICtR-package.R
\docType{package}
\name{PICtR-package}
\alias{PICtR}
\alias{PICtR-package}
\title{PICtR: Physically Interacting Cell (PIC) toolkit for R}
\description{
This is the computational workflow to analyse physically interacting cells (PICs) in flow cytometry data using R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/agSHaas/PICtR}
  \item Report bugs at \url{https://github.com/agSHaas/PICtR/issues}
}

}
\author{
\strong{Maintainer}: Viktoria Flore \email{viktoria.flore@bih-charite.de} (\href{https://orcid.org/0009-0001-3023-433X}{ORCID})

Authors:
\itemize{
  \item Lea Jopp-Saile \email{jopp-saile@protonmail.com} (\href{https://orcid.org/0000-0002-7032-3003}{ORCID})
}

}
\keyword{internal}
