% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{ratio_cluster_plot}
\alias{ratio_cluster_plot}
\title{Ratio cluster plot.}
\usage{
ratio_cluster_plot(
  obj,
  clusters = "seurat_clusters",
  ratio = "ratio_anno",
  assay = "FACS"
)
}
\arguments{
\item{obj}{The Seurat object.}

\item{clusters}{The meta.data column with the clustering solution.}

\item{ratio}{The meta.data column with the classification of cells based on the Otsu threshold of the FSC.A/FSC.H ratio.}

\item{assay}{The Seurat assay to use.}
}
\value{
None
}
\description{
Stacked bar plot for the proportion of cells below/above the Otsu threshold for the FSC.A/FSC.H ratio, in each cluster.
}
