% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{ratio_cluster_plot}
\alias{ratio_cluster_plot}
\title{Ratio cluster plot.}
\usage{
ratio_cluster_plot(
  obj,
  clusters = "seurat_clusters",
  ratio = "ratio_anno",
  assay = "FACS"
)
}
\arguments{
\item{obj}{The Seurat object.}

\item{clusters}{The string of the meta.data column with the clustering resolution to plot.}

\item{ratio}{The meta.data column with the classification of cells (ratio_high/ratio_low) determined using the FSC.A/FSC.H ratio and the determined threshold.}

\item{assay}{The Seurat assay to use (default FACS).}
}
\value{
None
}
\description{
Stacked bar plot of each cluster with the proportion of cells below/above the determined FSC.A/FSC.H ratio threshold.
}
