% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_dbt.R
\name{select_dbt}
\alias{select_dbt}
\title{Selection of clusters containing physically interacting cells.}
\usage{
select_dbt(
  obj,
  clusters = "seurat_clusters",
  ratio = "ratio_anno",
  ratio_high = "Ratio_high",
  assay = "FACS",
  quantile = 0.8,
  selected_clusters = "doublet_clusters"
)
}
\arguments{
\item{obj}{The Seurat object.}

\item{clusters}{The meta.data column with the clustering solution.}

\item{ratio}{The meta.data column with the classification of cells based on the Otsu threshold of the FSC.A/FSC.H ratio.}

\item{ratio_high}{The character string that indicates cells with a FSC.A/FSC.H ratio above the Otsu threshold within the meta.data column with the classification of cells based on the FSC.A/FSC.H ratio.}

\item{assay}{The Seurat assay.}

\item{quantile}{The desired percentile cutoff above which clusters are classified as physically interacting cell clusters.}

\item{selected_clusters}{Character vector for the misc slot in the Seurat object that will contain the selected physically interacting cell clusters.}
}
\value{
Seurat object
}
\description{
Selects the clusters that belong to a given percentile of the clusters with the largest physically interacting cells proportion based on
}
