% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_function.R
\name{sketch_wrapper}
\alias{sketch_wrapper}
\title{PICtR workflow wrapper}
\usage{
sketch_wrapper(
  channel = channel,
  meta_data = NULL,
  assay = "FACS",
  FSC.A = "FSC.A",
  FSC.H = "FSC.H",
  n_sketch_cells = 50000,
  n_components = "all",
  clst_algorithm = 1,
  resolution = c(0.5, 1, 2, 3, 4),
  min_clst_size = 100,
  meta.k = "auto",
  obj_name = "obj_sketched_non_projected",
  ratio = TRUE,
  thresholding_method = "otsu",
  hist_breaks = 2000,
  group_by = NULL,
  verbose = TRUE,
  BPcell_dir = NULL,
  overwrite = F,
  working_dir = getwd(),
  seed = 42
)
}
\arguments{
\item{channel}{A data frame with the dimensionality cells x flow cytometry parameters.}

\item{meta_data}{A data frame with meta_data for every cell (Default=NULL).}

\item{assay}{A character string with the name of the assay to be created (Default="FACS").}

\item{FSC.A}{The name (string) of the column containing the FSC.A scatter parameter (Default="FACS.A").}

\item{FSC.H}{The name (string) of the column containing the FSC.H scatter parameter (Default="FACS.H").}

\item{n_sketch_cells}{The number of cells to be subsampled by \code{\link[Seurat]{SketchData}} (Default=50000).}

\item{n_components}{The number of components computed and used during analysis. "all" or given as an integer (Default="all").}

\item{clst_algorithm}{Algorithm used for clustering. For Seurat's \code{\link[Seurat]{FindClusters}}: 1 = original Louvain algorithm; 2 = Louvain algorithm with multilevel refinement; 3 = SLM algorithm; 4 = Leiden algorithm). Leiden requires the leidenalg python.
Alternatively, "hdbscan" for Hierarchical Density-Based Spatial Clustering of Applications with Noise (HDBSCAN) on the UMAP embedding, see \code{\link[dbscan]{hdbscan}}),
"flowMeans" for non-parametric clustering and segmented-regression-based change point detection, see \code{\link[flowMeans]{flowMeans}},
or "flowSOM" for self-organizing maps, see \code{\link[Spectre]{run.flowsom}}, (Default: Louvain, 1)}

\item{resolution}{A numerical vector with the desired resolutions for clustering (Default: c(0.5,1,2,3,4)). Only used for Louvain, SLM, and Leiden clustering.}

\item{min_clst_size}{minimum cluster size for HDBSCAN. See details at \code{\link[dbscan]{hdbscan}} (Default: 100).}

\item{meta.k}{meta k for FlowSOM clustering, see \code{\link[Spectre]{run.flowsom}} (Default: "auto").}

\item{obj_name}{The name used for storing the Seurat object (character).}

\item{ratio}{Boolean variable. If TRUE the FSC ratio (FSC.A/FSC.H) will be calculated.}

\item{thresholding_method}{Method for thresholding. One of "otsu", "triangle", "kmeans", or any method from the autothresholdr package. For details see ?calculateThreshold() (Default = "otsu").}

\item{hist_breaks}{Number of histogram breaks for methods that rely on histograms. Should be >100 (Default: 2000).}

\item{group_by}{Optional grouping parameter to calculate the FSC ratio (FSC.A/FSC.H) thresholding method for.}

\item{verbose}{Verbosity (Boolean).}

\item{BPcell_dir}{Optional directory with the counts matrix for \code{\link[BPCells]{open_matrix_dir}}.}

\item{overwrite}{Overwrite existing BPCells directory? (Default: FALSE)}

\item{working_dir}{Directory path to be used as working directory (character string).}

\item{seed}{Seed for reproducibility (Default: 42).}
}
\value{
Seurat object.
}
\description{
Characterizes cells into ratio_high and ratio_low cells for physically interacting cell analysis,
samples a representative subset of cells using \code{\link[Seurat]{SketchData}} and runs the standard Seurat analysis workflow.
}
\examples{
obj <- sketch_wrapper(channel = demo_lcmv,
                      meta_data = demo_lcmv,
                      n_sketch_cells = 5000,
                      clst_algorithm = 1,
                      ratio = TRUE,
                      thresholding_method = "otsu", 
                      working_dir = tempdir())

}
\references{
Hao et al. Dictionary learning for integrative, multimodal and scalable single-cell analysis. Nature Biotechnology (2023). \doi{10.1038/s41587-023-01767-y}

Van Gassen S et al. (2015) FlowSOM: Using self-organizing maps for visualization and interpretation of cytometry data. Cytom Part J Int Soc Anal Cytol 87: 636-645. \doi{10.1002/cyto.a.22625}.

Aghaeepour, N., Nikolic, R., Hoos, H.H. and Brinkman, R.R. (2011), Rapid cell population identification in flow cytometry data. Cytometry, 79A: 6-13. \doi{10.1002/cyto.a.21007}.

Campello, R.J.G.B., Moulavi, D., Sander, J. (2013). Density-Based Clustering Based on Hierarchical Density Estimates. In: Pei, J., Tseng, V.S., Cao, L., Motoda, H., Xu, G. (eds) Advances in Knowledge Discovery and Data Mining. PAKDD 2013. Lecture Notes in Computer Science(), vol 7819. Springer, Berlin, Heidelberg. \doi{10.1007/978-3-642-37456-2_14}
}
