% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{split_plot_sketch}
\alias{split_plot_sketch}
\title{Split plot wrapper.}
\usage{
split_plot_sketch(obj, group_by = "seurat_clusters", split_by = "ratio_anno")
}
\arguments{
\item{obj}{The Seurat object.}

\item{group_by}{Parameter to group the plot by.}

\item{split_by}{Parameter to split the plot by.}
}
\value{
ggplot object
}
\description{
Dimensional reduction (UMAP) plot split by a given parameter.
Per default the returned split plots are rasterized using \code{\link[ggrastr]{geom_point_rast}}.
}
