% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{wrapper_for_plots}
\alias{wrapper_for_plots}
\title{Plot wrapper function.}
\usage{
wrapper_for_plots(
  obj = obj,
  feature_plot = TRUE,
  cluster_plot = TRUE,
  meta_list = list("ratio_anno"),
  feature_plot_colors = pals::parula(1000),
  ratio_plot_color = c(Ratio_low = "dodgerblue2", Ratio_high = "gold2"),
  reduction = "umap",
  alpha = 1,
  raster = TRUE,
  label_size = 3,
  cluster_handle = "sketch_snn_res",
  label_box = FALSE,
  assay = "sketch"
)
}
\arguments{
\item{obj}{The Seurat object.}

\item{feature_plot}{Plot \code{\link[Seurat]{FeaturePlot}}?}

\item{cluster_plot}{Plot \code{\link[Seurat]{DimPlot}} by cluster?}

\item{meta_list}{List of meta.data columns to color \code{\link[Seurat]{DimPlot}} by.}

\item{feature_plot_colors}{Color palette for \code{\link[Seurat]{FeaturePlot}}}

\item{ratio_plot_color}{Color palette for \code{\link{ratio_cluster_plot}}}

\item{reduction}{Reduction to use for plotting, for example UMAP.}

\item{alpha}{Alpha value for plotting.}

\item{raster}{Rasterize plots?}

\item{label_size}{Size of the labels.}

\item{cluster_handle}{Prefix for the clustering solutions in the meta.data slot.}

\item{label_box}{Plot boxes around labels?}

\item{assay}{The Seurat assay to use.}
}
\value{
A list with the requested plots.
}
\description{
A wrapper function for plots that are often used during exploratory analysis within the \link[=https://satijalab.org/seurat/]{Seurat framework}. Includes \code{\link[Seurat]{FeaturePlot}}, \code{\link[Seurat]{DimPlot}} for different parameters, and \code{\link{ratio_cluster_plot}}.
}
